+++ api/logger/logger.go
@@ -3,11 +3,20 @@ package logger
 import (
 	"github.com/devtron-labs/devtron/internal/middleware"
 	"github.com/devtron-labs/devtron/pkg/user"
-	"io/ioutil"
+	"io"
 	"log"
 	"net/http"
 	"time"
 )
+
+type AuditLoggerDTO struct {
+	UrlPath         string    `json:"urlPath"`
+	UserEmail       string    `json:"userEmail"`
+	UpdatedOn       time.Time `json:"updatedOn"`
+	QueryParams     string    `json:"queryParams"`
+	ApiResponseCode int       `json:"apiResponseCode"`
+	RequestPayload  []byte    `json:"requestPayload"`
+}
 
 type LoggingMiddlewareImpl struct {
 	userService user.UserService
@@ -33,11 +42,11 @@ func (impl LoggingMiddlewareImpl) LoggingMiddleware(next http.Handler) http.Hand
 		token := r.Header.Get("token")
 		userEmail, err := impl.userService.GetEmailFromToken(token)
 		if err != nil {
-			log.Printf("user does not exists")
+			log.Printf("AUDIT_LOG: user does not exists")
 		}
-		body, err := ioutil.ReadAll(r.Body)
+		body, err := io.ReadAll(r.Body)
 		if err != nil {
-			log.Printf("error reading request body for urlPath: %s queryParams: %s userEmail: %s", r.URL.Path, r.URL.Query().Encode(), userEmail)
+			log.Printf("AUDIT_LOG: error reading request body for urlPath: %s queryParams: %s userEmail: %s", r.URL.Path, r.URL.Query().Encode(), userEmail)
 		}
 		auditLogDto := &AuditLoggerDTO{
 			UrlPath:         r.URL.Path,
@@ -49,4 +58,8 @@ func (impl LoggingMiddlewareImpl) LoggingMiddleware(next http.Handler) http.Hand
 		}
 		LogRequest(auditLogDto)
 	})
+}
+
+func LogRequest(auditLogDto *AuditLoggerDTO) {
+	log.Printf("AUDIT_LOG: urlPath: %s, queryParams: %s, requestPayload: %s,updatedBy: %s, updatedOn: %s, apiResponseCode: %d", auditLogDto.UrlPath, auditLogDto.QueryParams, auditLogDto.RequestPayload, auditLogDto.UserEmail, auditLogDto.UpdatedOn, auditLogDto.ApiResponseCode)
 }
